Decompose SequencerContext into focused hooks
Split the 1201-line SequencerContext into an
orchestrator plus four focused hooks:

- useFillMode (107 lines): fill state, F-key shortcut
- usePatternMode (198 lines): pattern mode state
  machine, temp mode, T-key shortcut
- useTrackConfig (505 lines): all track/mixer config
  actions, cycle count ref
- usePlayback (383 lines): audio engine integration,
  handleStep via computeStep, spacebar shortcut

SequencerContext.tsx shrinks to 356 lines (130 of
which are type definitions for the public API).

Also moves useSelection and useDragPaint to
src/app/hooks/ directory for consistency.

Each hook exposes a reset() method; clearAll in the
orchestrator composes all resets.

The public useSequencer() API ({state, actions, meta})
is completely unchanged.
