+++ src/app/StepGrid.tsx
@@ -9,8 +9,8 @@ import TrackRow from './TrackRow';
 import AccentRow from './AccentRow';
 import StepPopover from './StepPopover';
 import ErrorBoundary from './ErrorBoundary';
-import { useDragPaint } from './useDragPaint';
-import { useSelection } from './useSelection';
+import { useDragPaint } from './hooks/useDragPaint';
+import { useSelection } from './hooks/useSelection';
 import type { TrackId, TrackPattern } from './types';
 import { getPatternLength } from './types';
 import { TRIGGER_FLASH_MS } from './constants';
